back-ends/key_images_back/requirements.txt
flask 
requests
psycopg[binary]
django 
djangorestframework
google-auth
CacheControl
drf_yasg
timeloop
python-json-logger
gunicorn[gevent]

